feat(cmsis): Include MCU vendor's CMSIS headers
Replace peripheral structure and register definitions created directly through datasheets with CMSIS headers. This is because CMSIS headers contain definitions for everything that MCU contains and they are less prone to register level errors as they are an ARM standard and maintained by the MCU vendor. The linker script is replaced with the MCU vendor supplied linker script as it is also maintained by the MCU vendor and is sufficient for this project.